Updated layer-message-status to expose a bug
The bug on the current platform can be replicated by sending a message and then refreshing the page. The page will often then show either an empty string to the left of message time or 'sent', as is the case with the SDK.  

This bug is most likely caused by the fact that the last message sent is missing it's deliveryStatus and readStatus within the message on refresh.